requirement.txt
streamlit
langchain
chromadb
openai
python-dotenv
pypdf
